requirements.txt
gspread==5.12.4
google-auth==2.27.0
pandas==2.2.0
beautifulsoup4==4.12.3
streamlit==1.31.1
requests==2.31.0
fpdf==1.7.2
google-api-python-client==2.118.0
textblob==0.17.1
nltk==3.8.1
vaderSentiment==3.3.2
python-dotenv==1.0.1
